requirements.txt
albumentations==1.2.0
streamlit==1.4.0

numpy==1.20.2
#opencv==4.6.0
opencv-python==4.5.1.48
pillow==6.0.0
scikit-image==0.18.1
tqdm==4.46.1

torch==1.8.1+cpu
torchvision==0.13.1

scikit-learn==1.0.2
pip==22.3.1
dataclasses==0.8
attrs==22.1.0
zlib==1.2.13
wheel==0.38.4
openssl==1.1.1t
cryptography==39.0.1
tornado==6.2
jinja2==3.1.2
setuptools==65.6.3
typing_extensions==4.4.0
packaging==22.0
zipp==3.11.0
future==0.18.3
